backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.11.9
<<<<<<< HEAD
pydantic==2.5.0
anthropic==0.25.0
redis==5.0.1
upstash-redis==0.15.0
=======
pillow==10.1.0
numpy
opencv-python-headless
easyocr==1.7.0
anthropic==0.25.0
redis==5.0.1
upstash-redis==0.15.0
openai-whisper==20231117
torch
torchaudio
>>>>>>> 948a37ea296a26d67b78ce30e503ad388b399512
